Moved auth utilities and jwt generator to server folder
